Module test_data_generation::data::field

Structs

Field
Represents a Fact for a character in a sample data entity that has been analyzed

Enums

FieldType
This enum is used to specify the possible data types a Field can represent

Traits

ConvertToBytes
this trait is used to overload the set() function