Struct test_data_generation::data_sample_parser::DataSampleParser

pub struct DataSampleParser {
    pub issues: bool,
    // some fields omitted
}

Represents the Parser for sample data to be used

Fields
issues: bool

indicates if there were issues parsing and anlyzing the data sample

Methods

impl DataSampleParser

fn new() -> DataSampleParser

Constructs a new DataSampleParser

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    let dsp = DataSampleParser::new();
}

fn new_with(path: &'static str) -> DataSampleParser

Constructs a new DataSampleParser

Arguments

path: &'static str - The full path name (including the file name and extension) to the configuraiton file.

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    // param: the path to the configuration  file
    let dsp = DataSampleParser::new_with("/config/tdg.yaml");
}

fn analyze_csv_file(&mut self, path: &'static str) -> Result<i32, String>

This function analyzes sample data that is a csv formatted string and returns a boolean if successful.

Arguments

path: &'static str - The full path name of the csv formatted sample data file.

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    let mut dsp = DataSampleParser::new();
    
    assert_eq!(dsp.analyze_csv_file("./tests/samples/sample-01.csv").unwrap(),1);
}

fn demo_date(&self) -> String

This function generates date as strings using the a demo profile

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    let dsp = DataSampleParser::new();

    // generate some test data using the demo functions
    println!("generate date:{}", dsp.demo_date());
}

fn demo_person_name(&self) -> String

This function generates people's names as strings using the a demo profile

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    let dsp = DataSampleParser::new();

    // generate some test data using the demo functions
    println!("generate date:{}", dsp.demo_person_name());
}

fn generate_by_field_name(&mut self, field: String) -> String

This function generates test data for the specified field name.

Arguments

field: String - The name of the field (e.g.: firstname) the represents the profile to use when generating the test data.

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    let mut dsp = DataSampleParser::new();
    
    dsp.analyze_csv_file("./tests/samples/sample-01.csv").unwrap();
    println!("Generated data for first name {}",dsp.generate_by_field_name("firstname".to_string()));
}

fn generate_record(&mut self) -> Vec<String>

This function Vec of generates test data fields.

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    let mut dsp = DataSampleParser::new();
    
    dsp.analyze_csv_file("./tests/samples/sample-01.csv").unwrap();
    println!("Generated data record: {:?}",dsp.generate_record());
}

fn running_with_issues(&self) -> &bool

This function returns a boolean that indicates if the data sample parsing had issues

Example

extern crate test_data_generation;

use test_data_generation::data_sample_parser::DataSampleParser;
    
fn main() {
    // initalize a new DataSampelParser
    // param: the path to the configuration file is wrong
    let dsp = DataSampleParser::new_with("./target/debug/config/tdg.yaml");

    // generate some test data using the demo functions
    assert_eq!(dsp.running_with_issues(), &false);
}