Struct test_data_generation::profile::pattern_placeholder::PatternPlaceholder

pub struct PatternPlaceholder {
    pub placeholder: BTreeMap<String, char>,
}

Represents a placeholder for all the symbols used in pattern definitions

Fields
placeholder: BTreeMap<String, char>

A PatternPlaceholder must have a PlaceholderMap that contains all the symbols used to categorize characters by type (e.g.: Upper Case Vowel)

Methods

impl PatternPlaceholder

fn new() -> PatternPlaceholder

Constructs a new PatternPlaceholder

Example

extern crate test_data_generation;

use test_data_generation::profile::pattern_placeholder::PatternPlaceholder;
    
fn main() {
    let placeholder = PatternPlaceholder::new();
}

fn get(&self, key: &str) -> char

This function returns a pattern symbol that represents the type of character

Example

extern crate test_data_generation;

use test_data_generation::profile::pattern_placeholder::PatternPlaceholder;
 
fn main() {
    let placeholder =  PatternPlaceholder::new();
 
    println!("Upper case vowel symbol: {:?}", placeholder.get(&"VowelUpper".to_string()));
}